test(ivy): split out provider tests
